pyproject.toml
[project]
name = "lojex"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.1.7",
    "django-ninja>=1.3.0",
    "pydantic[email]>=2.10.6",
    "python-decouple>=3.8",
]
